tests/wpt/web-platform-tests/content-security-policy/object-src/object-src-url-blocked.html

<!DOCTYPE html>
<html>

<head>
    <meta http-equiv="Content-Security-Policy" content="object-src 'none'; script-src 'self' 'unsafe-inline';">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>

<body>
    <script>
      var t = async_test("Should block the object and fire a spv");
      window.addEventListener('securitypolicyviolation', t.step_func_done(function(e) {
        assert_equals(e.violatedDirective, "object-src");
      }));
    </script>

    <object type="image/png" data="/content-security-policy/support/pass.png"></object>
</body>

</html>
